requirements.txt
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
python-decouple==3.3
python-dotenv==0.14.0
python-http-client==3.3.1
sendgrid==6.4.6
six==1.15.0
starkbank-ecdsa==1.1.0
Werkzeug==1.0.1
